+++ CHANGELOG.md
@@ -125,3 +125,12 @@ Print message extension json header and desired tag in error log during publish.
 #2.4.0.3-RELEASE
 Note worthy:
 Remove app-test.properties file when package.
+
+#2.4.1-RELEASE
+Note worthy:
+
+fix:
+publish problem in producers at start up when lookup address and dcc remove control both applied.
+NSQExtNotSupportedException added for new error in youzan nsqd.
+
+feature: filter to skip message according to extension header key specified in NSQConfig for consumer.